done fast

+++ 0x03-python-data_structures/100-print_python_list_info.c
@@ -8,7 +8,7 @@
  */
 void print_python_list_info(PyObject *p)
 {
-	Py_ssize_t len, i;
+	long int len, i;
 	PyObject *item;
 
 	len = PyList_Size(p);
